Day 23 - Auth Added but need incomplete

+++ app/Http/Controllers/JobController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Job;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -46,6 +47,8 @@ class JobController extends Controller
     }
 
     public function update(Job $job){
+        Gate::authorize('edit', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
@@ -60,6 +63,7 @@ class JobController extends Controller
     }
 
     public function destroy(Job $job){
+        Gate::authorize('edit', $job);
         $job->delete();
         return redirect('/jobs');
     }
